Extracted Database interface
Extracted inner deserialization classes from `Database` into new class files.
Removed un-needed `Database.getGson()` calls and `Database`
field from `Search`; used client.getGson() instead.
Extracted public method signatures to Database interface.
Moved method implementations to new `DatabaseImpl` class.
Fixed `Response` import to be only for `com.cloudant.client.api.model`
package.
