+++ src/main/java/com/cloudant/client/api/ClientBuilder.java
@@ -4,6 +4,9 @@ import com.cloudant.client.api.model.Index;
 import com.cloudant.client.api.model.Permissions;
 import com.cloudant.client.api.model.Shard;
 import com.cloudant.client.api.views.Key;
+import com.cloudant.client.internal.util.IndexDeserializer;
+import com.cloudant.client.internal.util.SecurityDeserializer;
+import com.cloudant.client.internal.util.ShardDeserializer;
 import com.cloudant.client.org.lightcouch.CouchDbException;
 import com.cloudant.client.org.lightcouch.CouchDbProperties;
 import com.cloudant.http.HttpConnectionInterceptor;
