+++ src/main/java/com/cloudant/client/api/Database.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2015 IBM Corp. All rights reserved.
+ * Copyright (c) 2016 IBM Corp. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file
  * except in compliance with the License. You may obtain a copy of the License at
@@ -14,89 +14,26 @@
 
 package com.cloudant.client.api;
 
-import static com.cloudant.client.org.lightcouch.internal.CouchDbUtil.assertNotEmpty;
-import static com.cloudant.client.org.lightcouch.internal.CouchDbUtil.close;
-import static com.cloudant.client.org.lightcouch.internal.CouchDbUtil.createPost;
-import static com.cloudant.client.org.lightcouch.internal.CouchDbUtil.getAsString;
-import static com.cloudant.client.org.lightcouch.internal.CouchDbUtil.getResponse;
-import static com.cloudant.client.org.lightcouch.internal.CouchDbUtil.getResponseList;
-
 import com.cloudant.client.api.model.DbInfo;
 import com.cloudant.client.api.model.FindByIndexOptions;
 import com.cloudant.client.api.model.Index;
 import com.cloudant.client.api.model.IndexField;
-import com.cloudant.client.api.model.IndexField.SortOrder;
 import com.cloudant.client.api.model.Params;
 import com.cloudant.client.api.model.Permissions;
+import com.cloudant.client.api.model.Response;
 import com.cloudant.client.api.model.Shard;
 import com.cloudant.client.api.views.AllDocsRequestBuilder;
 import com.cloudant.client.api.views.ViewRequestBuilder;
-import com.cloudant.client.internal.DatabaseURIHelper;
-import com.cloudant.client.internal.URIBase;
-import com.cloudant.client.internal.views.AllDocsRequestBuilderImpl;
-import com.cloudant.client.internal.views.AllDocsRequestResponse;
-import com.cloudant.client.internal.views.ViewQueryParameters;
-import com.cloudant.client.org.lightcouch.Changes;
-import com.cloudant.client.org.lightcouch.CouchDatabase;
-import com.cloudant.client.org.lightcouch.CouchDbException;
 import com.cloudant.client.org.lightcouch.DocumentConflictException;
 import com.cloudant.client.org.lightcouch.NoDocumentException;
-import com.cloudant.client.org.lightcouch.Response;
-import com.cloudant.http.Http;
-import com.cloudant.http.HttpConnection;
-import com.google.gson.Gson;
-import com.google.gson.JsonArray;
-import com.google.gson.JsonDeserializationContext;
-import com.google.gson.JsonDeserializer;
-import com.google.gson.JsonElement;
-import com.google.gson.JsonObject;
-import com.google.gson.JsonParseException;
-import com.google.gson.JsonParser;
-import com.google.gson.JsonPrimitive;
-import com.google.gson.reflect.TypeToken;
 
 import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.io.Reader;
-import java.io.UnsupportedEncodingException;
-import java.lang.reflect.Type;
 import java.net.URI;
-import java.util.ArrayList;
 import java.util.EnumSet;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Map.Entry;
-import java.util.Set;
-import java.util.logging.Logger;
 
-/**
- * Contains a Database Public API implementation.
- *
- * Methods may throw a {@link NoDocumentException} if the database does not exist.
- *
- * @author Mario Briggs
- * @since 0.0.1
- */
-public class Database {
-
-    static final Logger log = Logger.getLogger(Database.class.getCanonicalName());
-    private CouchDatabase db;
-    private CloudantClient client;
-    private final URI apiV2DBSecurityURI;
-
-    /**
-     * @param client
-     * @param db
-     */
-    Database(CloudantClient client, CouchDatabase db) {
-        super();
-        this.client = client;
-        this.db = db;
-        apiV2DBSecurityURI = new URIBase(client.getBaseUri()).path("_api").path("v2").path("db")
-                .path(db.getDbName()).path("_security").build();
-    }
-
+public interface Database {
     /**
      * Set permissions for a user/apiKey on this database.
      * <p>
@@ -130,62 +67,7 @@ public class Database {
      * href="http://docs.cloudant.com/authorization.html#modifying-permissions">Modifying
      * permissions</a>
      */
-    public void setPermissions(String userNameorApikey, EnumSet<Permissions> permissions) {
-        assertNotEmpty(userNameorApikey, "userNameorApikey");
-        assertNotEmpty(permissions, "permissions");
-        final JsonArray jsonPermissions = new JsonArray();
-        for (Permissions s : permissions) {
-            final JsonPrimitive permission = new JsonPrimitive(s.toString());
-            jsonPermissions.add(permission);
-        }
-
-        // get existing permissions
-        JsonObject perms = getPermissionsObject();
-
-        // now set back
-        JsonElement elem = perms.get("cloudant");
-        if (elem == null) {
-            perms.addProperty("_id", "_security");
-            elem = new JsonObject();
-            perms.add("cloudant", elem);
-        }
-        elem.getAsJsonObject().add(userNameorApikey, jsonPermissions);
-
-        InputStream response = null;
-        HttpConnection put = Http.PUT(apiV2DBSecurityURI, "application/json");
-        put.setRequestBody(client.getGson().toJson(perms));
-        try {
-            response = client.couchDbClient.executeToInputStream(put);
-            String ok = getAsString(response, "ok");
-            if (!ok.equalsIgnoreCase("true")) {
-                //raise exception
-            }
-        } finally {
-            close(response);
-        }
-    }
-
-    /**
-     * @return /api/v2/db/$dbname/_security JSON data
-     * @throws UnsupportedOperationException if called on a database that does not provide the
-     *                                       Cloudant authorization API
-     */
-    private JsonObject getPermissionsObject() {
-        try {
-            return client.couchDbClient.get(apiV2DBSecurityURI, JsonObject.class);
-        } catch (CouchDbException exception) {
-            //currently we can't inspect the HttpResponse code
-            //being in this catch block means it was not a 20x code
-            //look for the "bad request" that implies the endpoint is not supported
-            if (exception.getMessage().toLowerCase().contains("bad request")) {
-                throw new UnsupportedOperationException("The methods getPermissions and " +
-                        "setPermissions are not supported for this database, consider using the " +
-                        "/db/_security endpoint.");
-            } else {
-                throw exception;
-            }
-        }
-    }
+    void setPermissions(String userNameorApikey, EnumSet<Permissions> permissions);
 
     /**
      * Returns the Permissions of this database.
@@ -205,11 +87,7 @@ public class Database {
      * href="http://docs.cloudant.com/authorization.html#viewing-permissions">Viewing
      * permissions</a>
      */
-    public Map<String, EnumSet<Permissions>> getPermissions() {
-        JsonObject perms = getPermissionsObject();
-        return client.getGson().getAdapter(new TypeToken<Map<String, EnumSet<Permissions>>>() {
-        }).fromJsonTree(perms);
-    }
+    Map<String, EnumSet<Permissions>> getPermissions();
 
     /**
      * Get info about the shards in the database.
@@ -218,18 +96,7 @@ public class Database {
      * @see <a target="_blank"
      * href="https://docs.cloudant.com/advanced.html#get-/$db/_shards">_shards</a>
      */
-    public List<Shard> getShards() {
-        InputStream response = null;
-        try {
-            response = client.couchDbClient.get(new DatabaseURIHelper(db.getDBUri()).path("_shards")
-                    .build());
-            return getResponseList(response, client.getGson(),
-                    new TypeToken<List<Shard>>() {
-                    }.getType());
-        } finally {
-            close(response);
-        }
-    }
+    List<Shard> getShards();
 
     /**
      * Get info about the shard a document belongs to.
@@ -239,12 +106,7 @@ public class Database {
      * @see <a target="_blank"
      * href="https://docs.cloudant.com/advanced.html#get-/$db/_shards">_shards</a>
      */
-    public Shard getShard(String docId) {
-        assertNotEmpty(docId, "docId");
-        return client.couchDbClient.get(new DatabaseURIHelper(db.getDBUri()).path("_shards")
-                        .path(docId).build(),
-                Shard.class);
-    }
+    Shard getShard(String docId);
 
     /**
      * Create a new index
@@ -273,11 +135,7 @@ public class Database {
      * @param indexType     optional, type of index (only "json" as of now)
      * @param fields        array of fields in the index
      */
-    public void createIndex(String indexName, String designDocName, String indexType,
-                            IndexField[] fields) {
-        JsonObject indexDefn = getIndexDefinition(indexName, designDocName, indexType, fields);
-        createIndex(indexDefn.toString());
-    }
+    void createIndex(String indexName, String designDocName, String indexType, IndexField[] fields);
 
     /**
      * Create a new index from a JSON string
@@ -288,23 +146,7 @@ public class Database {
      * href="http://docs.cloudant.com/api/cloudant-query.html#creating-a-new-index">
      * index definition</a>
      */
-    public void createIndex(String indexDefinition) {
-        assertNotEmpty(indexDefinition, "indexDefinition");
-        InputStream putresp = null;
-        URI uri = new DatabaseURIHelper(db.getDBUri()).path("_index").build();
-        try {
-            putresp = client.couchDbClient.executeToInputStream(createPost(uri, indexDefinition,
-                    "application/json"));
-            String result = getAsString(putresp, "result");
-            if (result.equalsIgnoreCase("created")) {
-                log.info(String.format("Created Index: '%s'", indexDefinition));
-            } else {
-                log.warning(String.format("Index already exists : '%s'", indexDefinition));
-            }
-        } finally {
-            close(putresp);
-        }
-    }
+    void createIndex(String indexDefinition);
 
     /**
      * Find documents using an index
@@ -319,9 +161,7 @@ public class Database {
      * href="http://docs.cloudant.com/api/cloudant-query.html#cloudant-query-selectors">
      * selector syntax</a>
      */
-    public <T> List<T> findByIndex(String selectorJson, Class<T> classOfT) {
-        return findByIndex(selectorJson, classOfT, new FindByIndexOptions());
-    }
+    <T> List<T> findByIndex(String selectorJson, Class<T> classOfT);
 
     /**
      * Find documents using an index
@@ -351,35 +191,7 @@ public class Database {
      * href="http://docs.cloudant.com/api/cloudant-query.html#cloudant-query-selectors">
      * selector syntax</a>
      */
-    public <T> List<T> findByIndex(String selectorJson, Class<T> classOfT, FindByIndexOptions
-            options) {
-        assertNotEmpty(selectorJson, "selectorJson");
-        assertNotEmpty(options, "options");
-
-        URI uri = new DatabaseURIHelper(db.getDBUri()).path("_find").build();
-        JsonObject body = getFindByIndexBody(selectorJson, options);
-        InputStream stream = null;
-        try {
-            stream = client.couchDbClient.executeToInputStream(createPost(uri, body.toString(),
-                    "application/json"));
-            Reader reader = new InputStreamReader(stream, "UTF-8");
-            JsonArray jsonArray = new JsonParser().parse(reader)
-                    .getAsJsonObject().getAsJsonArray("docs");
-            List<T> list = new ArrayList<T>();
-            for (JsonElement jsonElem : jsonArray) {
-                JsonElement elem = jsonElem.getAsJsonObject();
-                T t = client.getGson().fromJson(elem, classOfT);
-                list.add(t);
-            }
-            return list;
-        } catch (UnsupportedEncodingException e) {
-            // This should never happen as every implementation of the java platform is required
-            // to support UTF-8.
-            throw new RuntimeException(e);
-        } finally {
-            close(stream);
-        }
-    }
+    <T> List<T> findByIndex(String selectorJson, Class<T> classOfT, FindByIndexOptions options);
 
     /**
      * List all indices
@@ -392,18 +204,7 @@ public class Database {
      *
      * @return List of Index objects
      */
-    public List<Index> listIndices() {
-        InputStream response = null;
-        try {
-            URI uri = new DatabaseURIHelper(db.getDBUri()).path("_index").build();
-            response = client.couchDbClient.get(uri);
-            return getResponseList(response, client.getGson(),
-                    new TypeToken<List<Index>>() {
-                    }.getType());
-        } finally {
-            close(response);
-        }
-    }
+    List<Index> listIndices();
 
     /**
      * Delete an index
@@ -411,20 +212,7 @@ public class Database {
      * @param indexName   name of the index
      * @param designDocId ID of the design doc
      */
-    public void deleteIndex(String indexName, String designDocId) {
-        assertNotEmpty(indexName, "indexName");
-        assertNotEmpty(designDocId, "designDocId");
-        URI uri = new DatabaseURIHelper(db.getDBUri()).path("_index").path(designDocId)
-                .path("json").path(indexName).build();
-        InputStream response = null;
-        try {
-            HttpConnection connection = Http.DELETE(uri);
-            response = client.couchDbClient.executeToInputStream(connection);
-            getResponse(response, Response.class, client.getGson());
-        } finally {
-            close(response);
-        }
-    }
+    void deleteIndex(String indexName, String designDocId);
 
     /**
      * Provides access to Cloudant <tt>Search</tt> APIs.
@@ -433,19 +221,15 @@ public class Database {
      * @return Search object for searching the index
      * @see <a target="_blank" href="https://docs.cloudant.com/search.html">Search</a>
      */
-    public Search search(String searchIndexId) {
-        return new Search(client, this, searchIndexId);
-    }
+    Search search(String searchIndexId);
 
     /**
      * Get a manager that has convenience methods for managing design documents.
-     * 
+     *
      * @return a {@link DesignDocumentManager} for this database
      * @see DesignDocumentManager
      */
-    public DesignDocumentManager getDesignDocumentManager() {
-        return new DesignDocumentManager(this);
-    }
+    DesignDocumentManager getDesignDocumentManager();
 
     /**
      * @param designDoc containing the view
@@ -455,9 +239,7 @@ public class Database {
      * @see <a target="_blank"
      * href="https://docs.cloudant.com/creating_views.html#using-views">Using views</a>
      */
-    public ViewRequestBuilder getViewRequestBuilder(String designDoc, String viewName) {
-        return new ViewRequestBuilder(client, this, designDoc, viewName);
-    }
+    ViewRequestBuilder getViewRequestBuilder(String designDoc, String viewName);
 
     /**
      * Build a request for the _all_docs endpoint.
@@ -472,34 +254,21 @@ public class Database {
      *
      * @return a request builder for the _all_docs endpoint of this database
      */
-    public AllDocsRequestBuilder getAllDocsRequestBuilder() {
-        return new AllDocsRequestBuilderImpl(new ViewQueryParameters<String,
-                AllDocsRequestResponse.Revision>(client, this, "", "", String.class,
-                AllDocsRequestResponse.Revision.class) {
-            protected DatabaseURIHelper getViewURIBuilder() {
-                return new DatabaseURIHelper(db.getDBUri()).path("_all_docs");
-            }
-        });
-    }
+    AllDocsRequestBuilder getAllDocsRequestBuilder();
 
     /**
      * Provides access for interacting with the changes feed.
      * <P>
-     * See the {@link com.cloudant.client.api.Changes} API for examples.
+     * See the {@link Changes} API for examples.
      * </P>
      *
      * @return a Changes object for using the changes feed
-     * @see com.cloudant.client.api.Changes
+     * @see Changes
      * @see <a target="_blank"
      * href="https://docs.cloudant.com/database.html#get-changes">Databases - get
      * changes</a>
      */
-    public com.cloudant.client.api.Changes changes() {
-        Changes couchDbChanges = db.changes();
-        com.cloudant.client.api.Changes changes = new com.cloudant.client.api.Changes
-                (couchDbChanges);
-        return changes;
-    }
+    Changes changes();
 
     /**
      * Retrieve the document with the specified ID from the database and deserialize to an
@@ -514,9 +283,7 @@ public class Database {
      * @see <a target="_blank" href="https://docs.cloudant.com/document.html#read">Documents -
      * read</a>
      */
-    public <T> T find(Class<T> classType, String id) {
-        return db.find(classType, id);
-    }
+    <T> T find(Class<T> classType, String id);
 
     /**
      * Retrieve the document with the specified ID from the database and deserialize to an
@@ -540,10 +307,7 @@ public class Database {
      * @see <a target="_blank" href="https://docs.cloudant.com/document.html#read">Documents -
      * read</a>
      */
-    public <T> T find(Class<T> classType, String id, Params params) {
-        assertNotEmpty(params, "params");
-        return db.find(classType, id, params.getInternalParams());
-    }
+    <T> T find(Class<T> classType, String id, Params params);
 
     /**
      * Retrieve the document with the specified ID at the specified revision from the database
@@ -564,9 +328,7 @@ public class Database {
      * @see <a target="_blank" href="https://docs.cloudant.com/document.html#read">Documents -
      * read</a>
      */
-    public <T> T find(Class<T> classType, String id, String rev) {
-        return db.find(classType, id, rev);
-    }
+    <T> T find(Class<T> classType, String id, String rev);
 
     /**
      * This method finds any document given a URI.
@@ -586,9 +348,7 @@ public class Database {
      * @param <T>       the type of Java object to return
      * @return an object of type T
      */
-    public <T> T findAny(Class<T> classType, String uri) {
-        return db.findAny(classType, uri);
-    }
+    <T> T findAny(Class<T> classType, String uri);
 
     /**
      * Finds the document with the specified document ID and returns it as an {@link InputStream}.
@@ -599,9 +359,7 @@ public class Database {
      * @throws NoDocumentException If the document is not found in the database.
      * @see #find(String, String)
      */
-    public InputStream find(String id) {
-        return db.find(id);
-    }
+    InputStream find(String id);
 
     /**
      * Finds the document with the specified document ID and revision and returns it as {@link
@@ -629,9 +387,7 @@ public class Database {
      * @see <a target="_blank" href="https://docs.cloudant.com/document.html#read">Documents -
      * read</a>
      */
-    public InputStream find(String id, String rev) {
-        return db.find(id, rev);
-    }
+    InputStream find(String id, String rev);
 
     /**
      * Checks if a document exists in the database.
@@ -639,9 +395,7 @@ public class Database {
      * @param id the document _id field
      * @return {@code true} if the document is found, {@code false} otherwise
      */
-    public boolean contains(String id) {
-        return db.contains(id);
-    }
+    boolean contains(String id);
 
     /**
      * Saves a document in the database.
@@ -684,12 +438,7 @@ public class Database {
      * @return {@link Response}
      * @throws DocumentConflictException If a conflict is detected during the save.
      */
-    public com.cloudant.client.api.model.Response save(Object object) {
-        Response couchDbResponse = db.save(object);
-        com.cloudant.client.api.model.Response response = new com.cloudant.client.api.model
-                .Response(couchDbResponse);
-        return response;
-    }
+    Response save(Object object);
 
     /**
      * Saves a document in the database similarly to {@link Database#save(Object)} but using a
@@ -704,12 +453,7 @@ public class Database {
      * href="https://docs.cloudant.com/document.html#quorum---writing-and-reading-data">
      * Documents - quorum</a>
      */
-    public com.cloudant.client.api.model.Response save(Object object, int writeQuorum) {
-        Response couchDbResponse = client.couchDbClient.put(getDBUri(), object, true, writeQuorum);
-        com.cloudant.client.api.model.Response response = new com.cloudant.client.api.model
-                .Response(couchDbResponse);
-        return response;
-    }
+    Response save(Object object, int writeQuorum);
 
     /**
      * Creates a document in the database using a HTTP {@code POST} request.
@@ -730,12 +474,7 @@ public class Database {
      * @see <a target="_blank"
      * href="https://docs.cloudant.com/document.html#documentCreate">Documents - create</a>
      */
-    public com.cloudant.client.api.model.Response post(Object object) {
-        Response couchDbResponse = db.post(object);
-        com.cloudant.client.api.model.Response response = new com.cloudant.client.api.model
-                .Response(couchDbResponse);
-        return response;
-    }
+    Response post(Object object);
 
     /**
      * Creates a document in the database similarly to {@link Database#post(Object)} but using a
@@ -749,22 +488,7 @@ public class Database {
      * href="https://docs.cloudant.com/document.html#quorum---writing-and-reading-data">
      * Documents - quorum</a>
      */
-    public com.cloudant.client.api.model.Response post(Object object, int writeQuorum) {
-        assertNotEmpty(object, "object");
-        InputStream response = null;
-        try {
-            URI uri = new DatabaseURIHelper(db.getDBUri()).query("w", writeQuorum).build();
-            response = client.couchDbClient.executeToInputStream(createPost(uri, client.getGson()
-                            .toJson(object),
-                    "application/json"));
-            Response couchDbResponse = getResponse(response, Response.class, client.getGson());
-            com.cloudant.client.api.model.Response cloudantResponse = new com.cloudant.client.api
-                    .model.Response(couchDbResponse);
-            return cloudantResponse;
-        } finally {
-            close(response);
-        }
-    }
+    Response post(Object object, int writeQuorum);
 
     /**
      * Updates an object in the database, the object must have the correct {@code _id} and
@@ -787,12 +511,7 @@ public class Database {
      * @see <a target="_blank" href="https://docs.cloudant.com/document.html#update">Documents -
      * update</a>
      */
-    public com.cloudant.client.api.model.Response update(Object object) {
-        Response couchDbResponse = db.update(object);
-        com.cloudant.client.api.model.Response response = new com.cloudant.client.api.model
-                .Response(couchDbResponse);
-        return response;
-    }
+    Response update(Object object);
 
     /**
      * Updates an object in the database similarly to {@link #update(Object)}, but specifying the
@@ -807,12 +526,7 @@ public class Database {
      * href="https://docs.cloudant.com/document.html#quorum---writing-and-reading-data">
      * Documents - quorum</a>
      */
-    public com.cloudant.client.api.model.Response update(Object object, int writeQuorum) {
-        Response couchDbResponse = client.couchDbClient.put(getDBUri(), object, false, writeQuorum);
-        com.cloudant.client.api.model.Response response = new com.cloudant.client.api.model
-                .Response(couchDbResponse);
-        return response;
-    }
+    Response update(Object object, int writeQuorum);
 
     /**
      * Removes a document from the database.
@@ -833,12 +547,7 @@ public class Database {
      * @see <a target="_blank" href="https://docs.cloudant.com/document.html#delete">Documents -
      * delete</a>
      */
-    public com.cloudant.client.api.model.Response remove(Object object) {
-        Response couchDbResponse = db.remove(object);
-        com.cloudant.client.api.model.Response response = new com.cloudant.client.api.model
-                .Response(couchDbResponse);
-        return response;
-    }
+    Response remove(Object object);
 
     /**
      * Removes the document from the database with the specified {@code _id} and {@code _rev}
@@ -857,12 +566,7 @@ public class Database {
      * @see <a target="_blank" href="https://docs.cloudant.com/document.html#delete">Documents -
      * delete</a>
      */
-    public com.cloudant.client.api.model.Response remove(String id, String rev) {
-        Response couchDbResponse = db.remove(id, rev);
-        com.cloudant.client.api.model.Response response = new com.cloudant.client.api.model
-                .Response(couchDbResponse);
-        return response;
-    }
+    Response remove(String id, String rev);
 
     /**
      * Uses the {@code _bulk_docs} endpoint to insert multiple documents into the database in a
@@ -885,17 +589,7 @@ public class Database {
      * @see <a target="_blank" href="https://docs.cloudant.com/document.html#bulk-operations">
      * Documents - bulk operations</a>
      */
-    public List<com.cloudant.client.api.model.Response> bulk(List<?> objects) {
-        List<Response> couchDbResponseList = db.bulk(objects, false);
-        List<com.cloudant.client.api.model.Response> cloudantResponseList = new ArrayList<com
-                .cloudant.client.api.model.Response>();
-        for (Response couchDbResponse : couchDbResponseList) {
-            com.cloudant.client.api.model.Response response = new com.cloudant.client.api.model
-                    .Response(couchDbResponse);
-            cloudantResponseList.add(response);
-        }
-        return cloudantResponseList;
-    }
+    List<Response> bulk(List<?> objects);
 
     /**
      * Creates an attachment from the specified InputStream and a new document with a generated
@@ -916,13 +610,7 @@ public class Database {
      * @return {@link Response}
      * @see <a target="_blank" href="https://docs.cloudant.com/attachments.html">Attachments</a>
      */
-    public com.cloudant.client.api.model.Response saveAttachment(InputStream in, String name,
-                                                                 String contentType) {
-        Response couchDbResponse = db.saveAttachment(in, name, contentType);
-        com.cloudant.client.api.model.Response response = new com.cloudant.client.api.model
-                .Response(couchDbResponse);
-        return response;
-    }
+    Response saveAttachment(InputStream in, String name, String contentType);
 
     /**
      * Creates or updates an attachment on the given document ID and revision.
@@ -954,14 +642,8 @@ public class Database {
      * @throws DocumentConflictException if the attachment cannot be saved because of a conflict
      * @see <a target="_blank" href="https://docs.cloudant.com/attachments.html">Attachments</a>
      */
-    public com.cloudant.client.api.model.Response saveAttachment(InputStream in, String name,
-                                                                 String contentType, String
-                                                                         docId, String docRev) {
-        Response couchDbResponse = db.saveAttachment(in, name, contentType, docId, docRev);
-        com.cloudant.client.api.model.Response response = new com.cloudant.client.api.model
-                .Response(couchDbResponse);
-        return response;
-    }
+    Response saveAttachment(InputStream in, String name, String contentType, String docId, String
+            docRev);
 
     /**
      * Invokes an Update Handler.
@@ -992,18 +674,12 @@ public class Database {
      * href="https://docs.cloudant.com/design_documents.html#update-handlers">
      * Design documents - update handlers</a>
      */
-    public String invokeUpdateHandler(String updateHandlerUri, String docId,
-                                      Params params) {
-        assertNotEmpty(params, "params");
-        return db.invokeUpdateHandler(updateHandlerUri, docId, params.getInternalParams(), null);
-    }
+    String invokeUpdateHandler(String updateHandlerUri, String docId, Params params);
 
     /**
      * @return The database URI.
      */
-    public URI getDBUri() {
-        return db.getDBUri();
-    }
+    URI getDBUri();
 
     /**
      * Get information about this database.
@@ -1012,10 +688,7 @@ public class Database {
      * @see <a target="_blank" href="https://docs.cloudant.com/database.html#read">Databases -
      * read</a>
      */
-    public DbInfo info() {
-        return client.couchDbClient.get(new DatabaseURIHelper(db.getDBUri()).getDatabaseUri(),
-                DbInfo.class);
-    }
+    DbInfo info();
 
     /**
      * Requests the database commits any recent changes to disk.
@@ -1025,203 +698,6 @@ public class Database {
      * CouchDB _ensure_full_commit
      * </a>
      */
-    public void ensureFullCommit() {
-        db.ensureFullCommit();
-    }
+    void ensureFullCommit();
 
-    // private helper methods
-
-    /**
-     * Form a create index json from parameters
-     */
-    private JsonObject getIndexDefinition(String indexName, String designDocName,
-                                      String indexType, IndexField[] fields) {
-        assertNotEmpty(fields, "index fields");
-        JsonObject indexObject = new JsonObject();
-        if (!(indexName == null || indexName.isEmpty())) {
-            indexObject.addProperty("name", indexName);
-        }
-        if (!(designDocName == null || designDocName.isEmpty())) {
-            indexObject.addProperty("ddoc", designDocName);
-        }
-        if (!(indexType == null || indexType.isEmpty())) {
-
-            indexObject.addProperty("type", indexType);
-        }
-
-        JsonArray fieldsArray = new JsonArray();
-        for (int i = 0; i < fields.length; i++) {
-            JsonObject fieldObject = new JsonObject();
-            fieldObject.addProperty(fields[i].getName(), fields[i].getOrder().toString());
-            fieldsArray.add(fieldObject);
-        }
-        JsonObject arrayOfFields = new JsonObject();
-        arrayOfFields.add("fields", fieldsArray);
-        indexObject.add("index", arrayOfFields);
-
-        return indexObject;
-    }
-
-    private JsonObject getFindByIndexBody(String selectorJson,
-                                      FindByIndexOptions options) {
-
-        JsonArray fieldsArray = new JsonArray();
-        if (options.getFields().size() > 0) {
-            for(String field : options.getFields()) {
-                JsonPrimitive jsonField = client.getGson().fromJson(field,
-                        JsonPrimitive.class);
-                fieldsArray.add(jsonField);
-            }
-        }
-
-        JsonArray sortArray = new JsonArray();
-        if (options.getSort().size() > 0) {
-
-            for(IndexField sort : options.getSort()) {
-                JsonObject sortObject = new JsonObject();
-                sortObject.addProperty(sort.getName(), sort.getOrder().toString());
-                sortArray.add(sortObject);
-            }
-        }
-
-        JsonObject indexObject = new JsonObject();
-
-        //parse and find if valid json issue #28
-        JsonObject selectorObject = null;
-        boolean isObject = true;
-        try {
-            selectorObject = getGson().fromJson(selectorJson, JsonObject.class);
-        } catch (JsonParseException e) {
-            isObject = false;
-        }
-
-        if(!isObject) {
-            if(selectorJson.startsWith("\"selector\"")) {
-                selectorJson = selectorJson.substring(selectorJson.indexOf(":") + 1,
-                        selectorJson.length()).trim();
-                selectorObject = getGson().fromJson(selectorJson, JsonObject.class);
-            } else {
-                throw new JsonParseException("selectorJson should be valid json or like " +
-                        "\"selector\": {...} ");
-            }
-        }
-
-        if(selectorObject.has("selector")) {
-            indexObject.add("selector", selectorObject.get("selector"));
-        } else {
-            indexObject.add("selector", selectorObject);
-        }
-
-        if(fieldsArray.size() > 0) {
-            indexObject.add("fields", fieldsArray);
-        }
-        if(sortArray.size() > 0) {
-            indexObject.add("sort", sortArray);
-        }
-        if (options.getLimit() != null) {
-            indexObject.addProperty("limit", options.getLimit());
-        }
-        if (options.getSkip() != null) {
-            indexObject.addProperty("skip", options.getSkip());
-        }
-        if (options.getReadQuorum() != null) {
-            indexObject.addProperty("r", options.getReadQuorum());
-        }
-        if (options.getUseIndex() != null) {
-            indexObject.add("use_index", getGson().fromJson(options.getUseIndex(),
-                    JsonArray.class));
-        }
-
-        return indexObject;
-    }
-
-    Gson getGson() {
-        return client.getGson();
-    }
-}
-
-class ShardDeserializer implements JsonDeserializer<List<Shard>> {
-
-
-    public List<Shard> deserialize(JsonElement json, Type typeOfT,
-                                   JsonDeserializationContext context) throws JsonParseException {
-
-        final List<Shard> shards = new ArrayList<Shard>();
-
-        final JsonObject jsonObject = json.getAsJsonObject();
-        Set<Map.Entry<String, JsonElement>> shardsObj = jsonObject.get("shards").getAsJsonObject
-                ().entrySet();
-
-        for (Entry<String, JsonElement> entry : shardsObj) {
-            String range = entry.getKey();
-            List<String> nodeNames = context.deserialize(entry.getValue(), new
-                    TypeToken<List<String>>() {
-                    }.getType());
-            shards.add(new Shard(range, nodeNames));
-        }
-
-        return shards;
-    }
-}
-
-
-class IndexDeserializer implements JsonDeserializer<List<Index>> {
-
-
-    public List<Index> deserialize(JsonElement json, Type typeOfT,
-                                   JsonDeserializationContext context) throws JsonParseException {
-
-        final List<Index> indices = new ArrayList<Index>();
-
-        final JsonObject jsonObject = json.getAsJsonObject();
-        JsonArray indArray = jsonObject.get("indexes").getAsJsonArray();
-        for (int i = 0; i < indArray.size(); i++) {
-            JsonObject ind = indArray.get(i).getAsJsonObject();
-            String ddoc = null;
-            if (!ind.get("ddoc").isJsonNull()) { // ddoc is optional
-                ddoc = ind.get("ddoc").getAsString();
-            }
-            Index idx = new Index(ddoc, ind.get("name").getAsString(),
-                    ind.get("type").getAsString());
-            JsonArray fldArray = ind.get("def").getAsJsonObject().get("fields").getAsJsonArray();
-            for (int j = 0; j < fldArray.size(); j++) {
-                Set<Map.Entry<String, JsonElement>> fld = fldArray.get(j).getAsJsonObject()
-                        .entrySet();
-                for (Entry<String, JsonElement> entry : fld) {
-                    idx.addIndexField(entry.getKey(),
-                            SortOrder.valueOf(entry.getValue().getAsString())
-                    );
-                }
-            }//end fldArray
-            indices.add(idx);
-
-        }// end indexes
-
-        return indices;
-    }
-}
-
-class SecurityDeserializer implements JsonDeserializer<Map<String, EnumSet<Permissions>>> {
-
-
-    public Map<String, EnumSet<Permissions>> deserialize(JsonElement json, Type typeOfT,
-                                                         JsonDeserializationContext context)
-            throws JsonParseException {
-
-        Map<String, EnumSet<Permissions>> perms = new HashMap<String, EnumSet<Permissions>>();
-        JsonElement elem = json.getAsJsonObject().get("cloudant");
-        if (elem == null) {
-            return perms;
-        }
-        Set<Map.Entry<String, JsonElement>> permList = elem.getAsJsonObject().entrySet();
-        for (Entry<String, JsonElement> entry : permList) {
-            String user = entry.getKey();
-            EnumSet<Permissions> p = context.deserialize(entry.getValue(), new
-                    TypeToken<EnumSet<Permissions>>() {
-                    }.getType());
-            perms.put(user, p);
-        }
-        return perms;
-
-    }
 }
